bread_mold_webapp/requirements.txt
Flask>=2.0.0
roboflow>=0.2.27
Pillow>=8.0.0
opencv-python-headless>=4.5.0
ultralytics>=8.0.0
setuptools<81
Flask
roboflow
Pillow
opencv-python-headless
ultralytics
